+++ src/core/modules.py
@@ -1,25 +1,3 @@
-#
-# modules.py
-#
-# The Pashmak Project
-# Copyright 2020-2021 parsa shahmaleki <parsampsh@gmail.com>
-#
-# This file is part of Pashmak.
-#
-# Pashmak is free software: you can redistribute it and/or modify
-# it under the terms of the GNU General Public License as published by
-# the Free Software Foundation, either version 3 of the License, or
-# (at your option) any later version.
-#
-# Pashmak is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with Pashmak.  If not, see <https://www.gnu.org/licenses/>.
-#########################################################################
-
 """ Internal modules """
 
 modules = {}
@@ -548,7 +526,31 @@ func printf($args)
         $file = $args[1]
     endif
     $file->write(str($obj))
-endfunc"""
+endfunc
+namespace pashmak
+    func zen
+        println('Zen of Pashmak\\n\\
+\\n\\
+The Zen of Pashmak is a collection of "guiding principles" for writing computer programs that influence the design of the Pashmak programming language. (Like zen of python). This fucking list is written by Mohammad Esmaeili.\\n\\
+\\n\\
+    Fucking syntax is better than beautiful syntax\\n\\
+    English is better than Finglish\\n\\
+    Lossless slow is better than loosing fast\\n\\
+    CatShit is better than DogShit\\n\\
+    DogShit is better than BullShit\\n\\
+    Chaos is better than peace\\n\\
+    Enthropy is better than order\\n\\
+    Crazy is better than logic\\n\\
+    Fun is better than boring\\n\\
+    Happy is better than sad\\n\\
+    Pashm is better than Hash\\n\\
+    While is better than Do-While\\n\\
+    Space is better than Tab\\n\\
+    Also tab is better than Space\\n\\
+    -> is better than .\\n\\
+    if-else is better than switch-case')
+    endfunc
+endns"""
 modules["sys"] = """#
 # sys.pashm
 #