+++ src/core/program.py
@@ -31,7 +31,7 @@ from . import helpers, version, modules, jit, parser, current_prog
 from .class_system import Class, ClassObject
 from .function import Function
 
-import hashlib, time, random, datetime, base64, json, http, http.cookies, http.server, http.client, http.cookiejar, socket, socketserver, math, pprint, subprocess, sqlite3, urllib, urllib.error, urllib.parse, urllib.request, urllib.response, urllib.robotparser
+import hashlib, time, random, datetime, base64, json, http, http.cookies, http.server, http.client, http.cookiejar, socket, socketserver, math, pprint, subprocess, sqlite3, urllib, urllib.error, urllib.parse, urllib.request, urllib.response, urllib.robotparser, platform
 
 class Program(helpers.Helpers):
     """ Pashmak program object """